moved gameLoop utility singleton to its own file. I will move the entire main loop and game instance to there soon
