+++ src/Main.java
@@ -115,14 +115,4 @@ public class Main {
 		System.out.println(identity.transform(vector).toString());		
 		**/
 	}
-
-}
-
-final class GameLoop{
-	private GameLoop(){};
-	public final static void update(Window window, InputController inputController, CameraController cameraController, double deltaTime){
-		cameraController.update((float)deltaTime);
-		inputController.reset();
-		window.update();
-	}
 }
